requirements.txt
flask
flask-cors
PyPDF2
pdfplumber
python-docx
docx2pdf
pdf2image
pytesseract
Pillow
reportlab
pandas
openpyxl
